# openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GeoStorm",
    "description": "Observability for AI-driven software discovery",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/runs": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "List Runs",
        "description": "Paginated list of runs for a project.",
        "operationId": "listRuns",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_RunResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Get Run",
        "description": "Run detail with aggregate perception score and competitors detected.",
        "operationId": "getRun",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/progress": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Stream Run Progress",
        "description": "Server-Sent Events stream for real-time run progress.",
        "operationId": "streamRunProgress",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/responses": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "List Responses",
        "description": "Paginated responses with nested mentions for a run.",
        "operationId": "listRunResponses",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ResponseItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/perception": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Get Perception",
        "description": "Time-series perception data for a project.",
        "operationId": "getPerception",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerceptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/trajectory": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Get Trajectory",
        "description": "Historical trajectory data showing week-over-week or day-over-day changes.",
        "operationId": "getTrajectory",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "daily",
                "weekly",
                "monthly"
              ],
              "type": "string",
              "default": "weekly",
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrajectoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/perception/breakdown": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Get Perception Breakdown",
        "description": "Per-term and per-provider perception breakdown from the latest scoring period.",
        "operationId": "getPerceptionBreakdown",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerceptionBreakdownResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_ResponseItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResponseItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[ResponseItem]"
      },
      "PaginatedResponse_RunResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[RunResponse]"
      },
      "PerceptionBreakdownByProvider": {
        "properties": {
          "provider_name": {
            "type": "string",
            "title": "Provider Name"
          },
          "recommendation_share": {
            "type": "number",
            "title": "Recommendation Share"
          },
          "position_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Avg"
          }
        },
        "type": "object",
        "required": [
          "provider_name",
          "recommendation_share"
        ],
        "title": "PerceptionBreakdownByProvider",
        "description": "Per-provider breakdown in perception data."
      },
      "PerceptionBreakdownByTerm": {
        "properties": {
          "term_id": {
            "type": "string",
            "title": "Term Id"
          },
          "term_name": {
            "type": "string",
            "title": "Term Name"
          },
          "recommendation_share": {
            "type": "number",
            "title": "Recommendation Share"
          },
          "position_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Avg"
          }
        },
        "type": "object",
        "required": [
          "term_id",
          "term_name",
          "recommendation_share"
        ],
        "title": "PerceptionBreakdownByTerm",
        "description": "Per-term breakdown in perception data."
      },
      "PerceptionBreakdownResponse": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "total_responses": {
            "type": "integer",
            "title": "Total Responses",
            "default": 0
          },
          "brand_mentions": {
            "type": "integer",
            "title": "Brand Mentions",
            "default": 0
          },
          "ranked_responses": {
            "type": "integer",
            "title": "Ranked Responses",
            "default": 0
          },
          "by_term": {
            "items": {
              "$ref": "#/components/schemas/PerceptionBreakdownByTerm"
            },
            "type": "array",
            "title": "By Term"
          },
          "by_provider": {
            "items": {
              "$ref": "#/components/schemas/PerceptionBreakdownByProvider"
            },
            "type": "array",
            "title": "By Provider"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "PerceptionBreakdownResponse",
        "description": "Perception breakdown with per-term and per-provider aggregated scores."
      },
      "PerceptionDataPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "overall_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Score"
          },
          "recommendation_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Share"
          },
          "position_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Avg"
          },
          "competitor_delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competitor Delta"
          },
          "trend_direction": {
            "$ref": "#/components/schemas/TrendDirection",
            "default": "stable"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "PerceptionDataPoint",
        "description": "Single data point in perception time-series."
      },
      "PerceptionResponse": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/PerceptionDataPoint"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "PerceptionResponse",
        "description": "Perception time-series data."
      },
...
      "ResponseItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "term_id": {
            "type": "string",
            "title": "Term Id"
          },
          "term_name": {
            "type": "string",
            "title": "Term Name"
          },
          "provider_name": {
            "type": "string",
            "title": "Provider Name"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "response_text": {
            "type": "string",
            "title": "Response Text"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "mentions": {
            "items": {
              "$ref": "#/components/schemas/MentionItem"
            },
            "type": "array",
            "title": "Mentions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "run_id",
          "term_id",
          "term_name",
          "provider_name",
          "model_name",
          "response_text",
          "created_at"
        ],
        "title": "ResponseItem",
        "description": "Response with nested mentions."
      },
      "RunDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "trigger_type": {
            "$ref": "#/components/schemas/TriggerType"
          },
          "triggered_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triggered By"
          },
          "total_queries": {
            "type": "integer",
            "title": "Total Queries",
            "default": 0
          },
          "completed_queries": {
            "type": "integer",
            "title": "Completed Queries",
            "default": 0
          },
          "failed_queries": {
            "type": "integer",
            "title": "Failed Queries",
            "default": 0
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "perception_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Perception Score"
          },
          "recommendation_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Share"
          },
          "competitors_detected": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Competitors Detected"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "status",
          "trigger_type",
          "created_at"
        ],
        "title": "RunDetailResponse",
        "description": "Run detail with perception data."
      },
      "RunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "trigger_type": {
            "$ref": "#/components/schemas/TriggerType"
          },
          "triggered_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triggered By"
          },
          "total_queries": {
            "type": "integer",
            "title": "Total Queries",
            "default": 0
          },
          "completed_queries": {
            "type": "integer",
            "title": "Completed Queries",
            "default": 0
          },
          "failed_queries": {
            "type": "integer",
            "title": "Failed Queries",
            "default": 0
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "status",
          "trigger_type",
          "created_at"
        ],
        "title": "RunResponse",
        "description": "Run summary."
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "description": "Status of a monitoring run."
      },
...
      "TrajectoryDataPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "recommendation_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Share"
          },
          "position_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Avg"
          },
          "competitor_delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competitor Delta"
          },
          "trend_direction": {
            "$ref": "#/components/schemas/TrendDirection",
            "default": "stable"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "TrajectoryDataPoint",
        "description": "Single data point in trajectory time-series."
      },
      "TrajectoryResponse": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/TrajectoryDataPoint"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "TrajectoryResponse",
        "description": "Historical trajectory data for a project."
      },
      "TrendDirection": {
        "type": "string",
        "enum": [
          "up",
          "down",
          "stable"
        ],
        "title": "TrendDirection",
        "description": "Trend direction indicators."
      },
      "TriggerType": {
        "type": "string",
        "enum": [
          "immediate",
          "scheduled",
          "manual"
        ],
        "title": "TriggerType",
        "description": "How a run was triggered."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}